requirements.txt
django
gunicorn
pymongo
pyJWT